==> Code/3_toy_model.Rmd <==
---
title: "Untitled"
output: html_document
---

Inspiration:
https://sahirbhatnagar.com/bayesian-stan
https://tysonbarrett.com/jekyll/update/2019/07/21/BayesianSims/

packages 

```{r}
library(pacman)
p_load(tidyverse, parallel, brms)
ncores = detectCores()
setwd("C:/Users/95/BayesianDecisionWeights/Code")
```

read simulated data 

```{r}

readeR <- function(name){
  
  df <- read_csv(name) %>%
    rename(p = probability,
         gammaT = gamma_true,
         deltaT = delta_true,
         gammaO = gamma_obs, 
         deltaO = delta_obs, 
         w = w_prob) %>%
  mutate(ID = as_factor(ID),
         item = as_factor(item))
  
}

df <- readeR("../Data/exp2_simulation.csv")
df %>% group_by(item) %>%
  summarize(test = mean(gammaO))
```

priors 

```{r}

(prior <- brms::get_prior(
  brms::bf(w ~ (deltaO * (p ** gammaO)) / 
                      (deltaO * (p ** gammaO) +
                         (1-p)**gammaO), 
               gammaO ~ 0 + item + (1|ID),
               deltaO ~ 0 + item + (1|ID),
               nl = TRUE),
  data = df, 
  family = gaussian()))

```

function for running models 

```{r}

brmsFun <- function(brm_prior, brm_df,
                    filename, iter){

  ## needs to be exp. & then we transfer back to log. 
  mod <- brms::brm(
  bf(w ~ (exp(logdeltaO) * (p ** gammaO)) / 
       (exp(logdeltaO) * (p **gammaO) +
          (1-p)**gammaO),
     gammaO ~ 0 + item + (1|ID),
     logdeltaO ~ 0 + item + (1|ID),
     nl = TRUE),
  data = brm_df, 
  prior = brm_prior,
  chains = 4,
  iter = iter,
  cores = 4,
  seed = 42,
  control = list(adapt_delta = 0.9),
  file = filename) 
  
}

```

specifying priors.

```{r}

prior1 <- c(prior(normal(0, 1), 
                  nlpar = "logdeltaO"),
            prior(normal(0.3, 0.5), 
                  nlpar = "gammaO"))

```

Run the model: 

```{r}
## for priors, see below: 
## (more than half an hour). 
fit <- brmsFun(brm_prior = prior1, 
                brm_df = df,
                filename = "../Data/fit_simulated",
                iter = 4000)

```
